fix(): allow non-object params to continue parsing
