+++ shared/jsx-storyboard/src/utils.ts
@@ -1,5 +1,16 @@
 import * as t from "@babel/types";
 
+const EXPRESSION_PREFIX_REG = /^\s*<%=?\s+/;
+const EXPRESSION_SUFFIX_REG = /\s+%>\s*$/;
+
+export function isExpressionString(value: unknown): value is string {
+  return (
+    typeof value === "string" &&
+    EXPRESSION_PREFIX_REG.test(value) &&
+    EXPRESSION_SUFFIX_REG.test(value)
+  );
+}
+
 export interface ValidateExpressionOptions {
   disallowArrowFunction?: boolean;
 }
